+++ crates/core/src/eval/functions/statistical/distributions_impl.rs
@@ -1488,341 +1488,5 @@ pub fn covar_fn(args: &[Value]) -> Value {
     Value::Number(cov)
 }
 
-// ---------------------------------------------------------------------------
-// Tests
-// ---------------------------------------------------------------------------
 #[cfg(test)]
-mod tests {
-    use super::*;
-
-    const EPS: f64 = 1e-4;
-
-    fn approx_eq(a: f64, b: f64) -> bool {
-        (a - b).abs() < EPS
-    }
-
-    fn num(n: f64) -> Value {
-        Value::Number(n)
-    }
-
-    fn bool_val(b: bool) -> Value {
-        Value::Bool(b)
-    }
-
-    fn nums_array(ns: &[f64]) -> Value {
-        Value::Array(ns.iter().map(|&n| Value::Number(n)).collect())
-    }
-
-    // -----------------------------------------------------------------------
-    // NORM.S.DIST
-    // -----------------------------------------------------------------------
-    #[test]
-    fn norm_s_dist_cdf_at_zero_is_half() {
-        let result = norm_s_dist_fn(&[num(0.0), bool_val(true)]);
-        match result {
-            Value::Number(v) => assert!(approx_eq(v, 0.5), "expected ~0.5, got {v}"),
-            other => panic!("expected Number, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn norm_s_dist_pdf_at_zero() {
-        let result = norm_s_dist_fn(&[num(0.0), bool_val(false)]);
-        match result {
-            Value::Number(v) => assert!(approx_eq(v, 0.3989), "expected ~0.3989, got {v}"),
-            other => panic!("expected Number, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn norm_s_dist_no_args_is_na() {
-        assert_eq!(norm_s_dist_fn(&[]), Value::Error(ErrorKind::NA));
-    }
-
-    // -----------------------------------------------------------------------
-    // NORM.DIST
-    // -----------------------------------------------------------------------
-    #[test]
-    fn norm_dist_standard_cdf_at_zero() {
-        // NORM.DIST(0, 0, 1, true) ≈ 0.5
-        let result = norm_dist_fn(&[num(0.0), num(0.0), num(1.0), bool_val(true)]);
-        match result {
-            Value::Number(v) => assert!(approx_eq(v, 0.5), "expected ~0.5, got {v}"),
-            other => panic!("expected Number, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn norm_dist_cdf_at_one_sigma() {
-        // NORM.DIST(1, 0, 1, true) ≈ 0.8413
-        let result = norm_dist_fn(&[num(1.0), num(0.0), num(1.0), bool_val(true)]);
-        match result {
-            Value::Number(v) => assert!(approx_eq(v, 0.8413), "expected ~0.8413, got {v}"),
-            other => panic!("expected Number, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn norm_dist_stdev_zero_is_num_error() {
-        let result = norm_dist_fn(&[num(0.0), num(0.0), num(0.0), bool_val(true)]);
-        assert_eq!(result, Value::Error(ErrorKind::Num));
-    }
-
-    #[test]
-    fn norm_dist_too_few_args_is_na() {
-        assert_eq!(norm_dist_fn(&[num(0.0), num(0.0)]), Value::Error(ErrorKind::NA));
-    }
-
-    // -----------------------------------------------------------------------
-    // NORM.S.INV
-    // -----------------------------------------------------------------------
-    #[test]
-    fn norm_s_inv_at_half_is_zero() {
-        let result = norm_s_inv_fn(&[num(0.5)]);
-        match result {
-            Value::Number(v) => assert!(approx_eq(v, 0.0), "expected ~0.0, got {v}"),
-            other => panic!("expected Number, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn norm_s_inv_at_0975_is_1_96() {
-        let result = norm_s_inv_fn(&[num(0.975)]);
-        match result {
-            Value::Number(v) => assert!(approx_eq(v, 1.96), "expected ~1.96, got {v}"),
-            other => panic!("expected Number, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn norm_s_inv_out_of_range_is_num_error() {
-        assert_eq!(norm_s_inv_fn(&[num(0.0)]), Value::Error(ErrorKind::Num));
-        assert_eq!(norm_s_inv_fn(&[num(1.0)]), Value::Error(ErrorKind::Num));
-    }
-
-    #[test]
-    fn norm_s_inv_no_args_is_na() {
-        assert_eq!(norm_s_inv_fn(&[]), Value::Error(ErrorKind::NA));
-    }
-
-    // -----------------------------------------------------------------------
-    // STANDARDIZE
-    // -----------------------------------------------------------------------
-    #[test]
-    fn standardize_basic() {
-        // STANDARDIZE(2, 1, 1) → 1.0
-        let result = standardize_fn(&[num(2.0), num(1.0), num(1.0)]);
-        assert_eq!(result, Value::Number(1.0));
-    }
-
-    #[test]
-    fn standardize_x_equals_mean_is_zero() {
-        let result = standardize_fn(&[num(5.0), num(5.0), num(2.0)]);
-        assert_eq!(result, Value::Number(0.0));
-    }
-
-    #[test]
-    fn standardize_stdev_zero_is_num_error() {
-        let result = standardize_fn(&[num(2.0), num(1.0), num(0.0)]);
-        assert_eq!(result, Value::Error(ErrorKind::Num));
-    }
-
-    #[test]
-    fn standardize_too_few_args_is_na() {
-        assert_eq!(standardize_fn(&[num(2.0)]), Value::Error(ErrorKind::NA));
-    }
-
-    // -----------------------------------------------------------------------
-    // CORREL / PEARSON
-    // -----------------------------------------------------------------------
-    #[test]
-    fn correl_perfect_positive() {
-        let result = correl_fn(&[nums_array(&[1.0, 2.0, 3.0]), nums_array(&[1.0, 2.0, 3.0])]);
-        match result {
-            Value::Number(v) => assert!(approx_eq(v, 1.0), "expected ~1.0, got {v}"),
-            other => panic!("expected Number, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn correl_perfect_negative() {
-        let result = correl_fn(&[nums_array(&[1.0, 2.0, 3.0]), nums_array(&[3.0, 2.0, 1.0])]);
-        match result {
-            Value::Number(v) => assert!(approx_eq(v, -1.0), "expected ~-1.0, got {v}"),
-            other => panic!("expected Number, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn correl_too_few_args_is_na() {
-        assert_eq!(correl_fn(&[nums_array(&[1.0, 2.0, 3.0])]), Value::Error(ErrorKind::NA));
-    }
-
-    #[test]
-    fn pearson_perfect_positive() {
-        let result = pearson_fn(&[nums_array(&[1.0, 2.0, 3.0]), nums_array(&[1.0, 2.0, 3.0])]);
-        match result {
-            Value::Number(v) => assert!(approx_eq(v, 1.0), "expected ~1.0, got {v}"),
-            other => panic!("expected Number, got {other:?}"),
-        }
-    }
-
-    // -----------------------------------------------------------------------
-    // POISSON / POISSON.DIST
-    // -----------------------------------------------------------------------
-    #[test]
-    fn poisson_pmf_at_zero_lambda_1() {
-        // POISSON(0, 1, false) ≈ e^-1 ≈ 0.3679
-        let result = poisson_fn(&[num(0.0), num(1.0), bool_val(false)]);
-        match result {
-            Value::Number(v) => assert!(approx_eq(v, 0.3679), "expected ~0.3679, got {v}"),
-            other => panic!("expected Number, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn poisson_cdf_at_zero_lambda_1() {
-        // POISSON(0, 1, true) ≈ e^-1 ≈ 0.3679
-        let result = poisson_fn(&[num(0.0), num(1.0), bool_val(true)]);
-        match result {
-            Value::Number(v) => assert!(approx_eq(v, 0.3679), "expected ~0.3679, got {v}"),
-            other => panic!("expected Number, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn poisson_too_few_args_is_na() {
-        assert_eq!(poisson_fn(&[num(0.0), num(1.0)]), Value::Error(ErrorKind::NA));
-    }
-
-    // -----------------------------------------------------------------------
-    // NORM.INV
-    // -----------------------------------------------------------------------
-    #[test]
-    fn norm_inv_at_half_is_mean() {
-        // NORM.INV(0.5, 3, 1) ≈ 3.0 (median = mean for normal distribution)
-        let result = norm_inv_fn(&[num(0.5), num(3.0), num(1.0)]);
-        match result {
-            Value::Number(v) => assert!(approx_eq(v, 3.0), "expected ~3.0, got {v}"),
-            other => panic!("expected Number, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn norm_inv_stdev_zero_is_num_error() {
-        let result = norm_inv_fn(&[num(0.5), num(0.0), num(0.0)]);
-        assert_eq!(result, Value::Error(ErrorKind::Num));
-    }
-
-    // -----------------------------------------------------------------------
-    // EXPON.DIST
-    // -----------------------------------------------------------------------
-    #[test]
-    fn expon_dist_cdf_at_zero_is_zero() {
-        // EXPON.DIST(0, 1, true) = 1 - e^0 = 0
-        let result = expon_dist_fn(&[num(0.0), num(1.0), bool_val(true)]);
-        assert_eq!(result, Value::Number(0.0));
-    }
-
-    #[test]
-    fn expon_dist_pdf_at_zero_equals_lambda() {
-        // EXPON.DIST(0, 2, false) = 2 * e^0 = 2
-        let result = expon_dist_fn(&[num(0.0), num(2.0), bool_val(false)]);
-        assert_eq!(result, Value::Number(2.0));
-    }
-
-    #[test]
-    fn expon_dist_lambda_zero_is_num_error() {
-        let result = expon_dist_fn(&[num(1.0), num(0.0), bool_val(true)]);
-        assert_eq!(result, Value::Error(ErrorKind::Num));
-    }
-
-    // -----------------------------------------------------------------------
-    // BINOM.DIST
-    // -----------------------------------------------------------------------
-    #[test]
-    fn binom_dist_pmf_certain_success() {
-        // BINOM.DIST(1, 1, 1.0, false) = 1.0 (k=n, p=1)
-        let result = binom_dist_fn(&[num(1.0), num(1.0), num(1.0), bool_val(false)]);
-        assert_eq!(result, Value::Number(1.0));
-    }
-
-    #[test]
-    fn binom_dist_cdf_all_successes() {
-        // BINOM.DIST(3, 3, 0.5, true) = 1.0 (CDF at max k = n)
-        let result = binom_dist_fn(&[num(3.0), num(3.0), num(0.5), bool_val(true)]);
-        assert_eq!(result, Value::Number(1.0));
-    }
-
-    #[test]
-    fn binom_dist_k_gt_n_is_num_error() {
-        let result = binom_dist_fn(&[num(4.0), num(3.0), num(0.5), bool_val(false)]);
-        assert_eq!(result, Value::Error(ErrorKind::Num));
-    }
-
-    // -----------------------------------------------------------------------
-    // FISHER / FISHERINV
-    // -----------------------------------------------------------------------
-    #[test]
-    fn fisher_at_zero_is_zero() {
-        let result = fisher_fn(&[num(0.0)]);
-        assert_eq!(result, Value::Number(0.0));
-    }
-
-    #[test]
-    fn fisher_out_of_range_is_num_error() {
-        assert_eq!(fisher_fn(&[num(1.0)]), Value::Error(ErrorKind::Num));
-        assert_eq!(fisher_fn(&[num(-1.0)]), Value::Error(ErrorKind::Num));
-    }
-
-    #[test]
-    fn fisher_inv_at_zero_is_zero() {
-        let result = fisher_inv_fn(&[num(0.0)]);
-        assert_eq!(result, Value::Number(0.0));
-    }
-
-    #[test]
-    fn fisher_roundtrip() {
-        // fisher_inv(fisher(0.5)) ≈ 0.5
-        let z = fisher_fn(&[num(0.5)]);
-        let x = match z {
-            Value::Number(v) => fisher_inv_fn(&[num(v)]),
-            other => panic!("unexpected {other:?}"),
-        };
-        match x {
-            Value::Number(v) => assert!(approx_eq(v, 0.5), "expected ~0.5, got {v}"),
-            other => panic!("expected Number, got {other:?}"),
-        }
-    }
-
-    // -----------------------------------------------------------------------
-    // GAUSS
-    // -----------------------------------------------------------------------
-    #[test]
-    fn gauss_at_zero_is_zero() {
-        // GAUSS(0) = norm_s_cdf(0) - 0.5 ≈ 0
-        let result = gauss_fn(&[num(0.0)]);
-        match result {
-            Value::Number(v) => assert!(approx_eq(v, 0.0), "expected ~0.0, got {v}"),
-            other => panic!("expected Number, got {other:?}"),
-        }
-    }
-
-    // -----------------------------------------------------------------------
-    // COVAR
-    // -----------------------------------------------------------------------
-    #[test]
-    fn covar_identical_arrays() {
-        // COVAR([1,2,3], [1,2,3]) = variance = 2/3
-        let result = covar_fn(&[nums_array(&[1.0, 2.0, 3.0]), nums_array(&[1.0, 2.0, 3.0])]);
-        match result {
-            Value::Number(v) => assert!(approx_eq(v, 2.0 / 3.0), "expected ~0.6667, got {v}"),
-            other => panic!("expected Number, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn covar_too_few_args_is_na() {
-        assert_eq!(covar_fn(&[nums_array(&[1.0, 2.0])]), Value::Error(ErrorKind::NA));
-    }
-}
+mod tests;
